requirements.txt
absl-py==0.14.0
addict==2.4.0
basicsr==1.3.4.3
cachetools==4.2.2
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1613247308275/work
charset-normalizer==2.0.6
colorama==0.4.4
cycler==0.10.0
Cython==0.29.22
facexlib==0.2.1.0
filterpy==1.4.5
future==0.18.2
gfpgan==0.2.1
google-auth==2.2.1
google-auth-oauthlib==0.4.6
grpcio==1.41.0
idna==3.2
imageio==2.9.0
importlib-metadata==4.8.1
kiwisolver==1.3.1
llvmlite==0.37.0
lmdb==1.2.1
Markdown==3.3.4
matplotlib==3.4.0
mkl-fft==1.3.0
mkl-random==1.0.4
mkl-service==2.3.0
networkx==2.6.3
numba==0.54.0
numpy @ file:///C:/ci/numpy_and_numpy_base_1603468620949/work
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.3.56
Pillow @ file:///C:/ci/pillow_1615224334542/work
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools-windows==2.0.0.2
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyparsing==2.4.7
pytesseract==0.3.8
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
# Editable install with no version control (realesrgan==0.2.2.4)
-e c:\users\nr\deepvision
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scipy==1.7.1
six @ file:///C:/ci/six_1605205426665/work
tb-nightly==2.7.0a20210927
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tifffile==2021.8.30
torch==1.8.1
torchsummary==1.5.1
torchvision==0.9.1
tqdm==4.62.3
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3==1.26.7
Werkzeug==2.0.1
wincertstore==0.2
yapf==0.31.0
zipp==3.5.1
